kan redigera bokade kursdeltagare som admin

+++ Varldsklass.Web/Controllers/CourseController.cs
@@ -8,6 +8,7 @@ using Varldsklass.Domain.Entities;
 using Varldsklass.Domain.Repositories;
 using Varldsklass.Domain.Repositories.Abstract;
 using System.Data.Entity;
+using Varldsklass.Web.ViewModels;
 
 namespace Varldsklass.Web.Controllers
 {
@@ -16,17 +17,33 @@ namespace Varldsklass.Web.Controllers
         private IRepository<Event> _eventRepo;
         private IRepository<Post> _postRepo;
         private IRepository<Category> _categoryRepo;
+        private IRepository<Attendant> _attendantRepo;
+        private IAccountRepository _accountRepo;
 
-        public CourseController(IRepository<Post> repo, IRepository<Category> category, IRepository<Event> Event)
+        public CourseController(IRepository<Post> repo, IRepository<Category> category, IRepository<Event> Event, IRepository<Attendant> attendantRepo, IAccountRepository accountRepo)
         {
             _eventRepo = Event;
             _postRepo = repo;
             _categoryRepo = category;
+            _attendantRepo = attendantRepo;
+            _accountRepo = accountRepo;
         }
 
+        private bool NotAllowedHere()
+        {
+            string email = User.Identity.Name;
+            return (_accountRepo.FindAll(u => u.Email == email && u.Administrator).Count() < 1);
+        }
+
+        private ActionResult RedirectAway()
+        {
+            return RedirectToAction("Index", "Home");
+        }
 
         public ActionResult Index()
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             AddCourseViewModel vm = new AddCourseViewModel();
             vm.Posts = _postRepo.FindAll().Where(p => p.postType == (int)Post.PostType.Course).ToList();
             vm.Categories = _categoryRepo.FindAll().ToList();
@@ -38,19 +55,51 @@ namespace Varldsklass.Web.Controllers
          * -------------------------------------*/
         public ActionResult Course(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             var posts = _postRepo.FindAll();
             return View(posts.Where(p => p.ID == id).Include(p => p.Events).FirstOrDefault());
         }
 
         public ActionResult EventAttendants(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             Event anEvent = _eventRepo.FindByID(id);
             return View(anEvent);
         }
 
+        [HttpGet]
+        public ActionResult EditAttendant(int id)
+        {
+            if (NotAllowedHere()) return RedirectAway();
+
+            Attendant attendant = _attendantRepo.FindByID(id);
+            return View(attendant);
+        }
+
+        [HttpPost]
+        public ActionResult EditAttendant(Attendant attendant)
+        {
+            if (NotAllowedHere()) return RedirectAway();
+
+            _attendantRepo.Save(attendant);
+            return RedirectToAction("EventAttendants", new { id = attendant.EventID });
+        }
+
+        public ActionResult DeleteAttendant(int id)
+        {
+            if (NotAllowedHere()) return RedirectAway();
+
+            int eventId = _attendantRepo.FindByID(id).Event.ID;
+            _attendantRepo.Delete( _attendantRepo.FindByID(id) );
+            return RedirectToAction("EventAttendants", new { id = eventId });
+        }
 
         public ActionResult EditCourse(int id = 0)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             if (id != 0)
             {
                 AddCourseViewModel vmCategory = new AddCourseViewModel();
@@ -73,6 +122,8 @@ namespace Varldsklass.Web.Controllers
 
         public ActionResult CourseInfo(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             var info = _postRepo.FindByID(id);
             return View(info);
         }
@@ -80,6 +131,8 @@ namespace Varldsklass.Web.Controllers
         /* ---- Front View ---- */
         public ActionResult CourseSingle(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             var posts = _postRepo.FindAll().Where(p => p.ID == id).Include(p => p.Events).FirstOrDefault();
 
             return View(posts);
@@ -88,7 +141,8 @@ namespace Varldsklass.Web.Controllers
         [HttpPost]
         public ActionResult SaveCourse(Post post, FormCollection postedForm)
         {
-            
+            if (NotAllowedHere()) return RedirectAway();
+
             if (ModelState.IsValid)
             {
                 var listOfCategoryIDs = postedForm["name"];
@@ -124,6 +178,8 @@ namespace Varldsklass.Web.Controllers
 
         public ActionResult DeleteCourse(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             _postRepo.Delete(_postRepo.FindByID(id));
             return RedirectToAction("Index");
         }
@@ -134,6 +190,8 @@ namespace Varldsklass.Web.Controllers
         * -------------------------------------*/
         public ActionResult AddEvent(int id = 0)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             ModelState.Clear();
             Event Event = new Event();
             Event.PostID = id;
@@ -145,6 +203,7 @@ namespace Varldsklass.Web.Controllers
 
         public ActionResult EditEvent(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
 
             var Event = _eventRepo.FindByID(id);
             return View("AddEvent", Event);
@@ -153,6 +212,8 @@ namespace Varldsklass.Web.Controllers
         [HttpPost]
         public ActionResult SaveEvent(Event Event)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             if (ModelState.IsValid)
             {
                 _eventRepo.Save(Event);
@@ -170,6 +231,8 @@ namespace Varldsklass.Web.Controllers
 
         public ActionResult DeleteEvent(int id, int course)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             _eventRepo.Delete(_eventRepo.FindByID(id));
             return RedirectToAction("Course", new { id = course });
         }
@@ -177,6 +240,8 @@ namespace Varldsklass.Web.Controllers
         /* ---- Front View ---- */
         public ActionResult EventSingle(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             var Event = _eventRepo.FindAll().Where(p => p.ID == id).Include(p => p.Post).FirstOrDefault();
 
             return View(Event);
@@ -186,6 +251,8 @@ namespace Varldsklass.Web.Controllers
         * -------------------------------------*/
         public ActionResult EditCategory(int id = 0)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             if (id != 0) {
                 return View("AddCategory", _categoryRepo.FindByID(id));
             } else {
@@ -198,6 +265,8 @@ namespace Varldsklass.Web.Controllers
         [HttpPost]
         public ActionResult SaveCategory(Category Category)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             if (ModelState.IsValid)
             {
                 _categoryRepo.Save(Category);
@@ -215,6 +284,8 @@ namespace Varldsklass.Web.Controllers
 
         public ActionResult DeleteCategory(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             _categoryRepo.Delete(_categoryRepo.FindByID(id));
             return RedirectToAction("Index");
         }
@@ -222,6 +293,8 @@ namespace Varldsklass.Web.Controllers
         /* ---- Front View ---- */
         public ActionResult Category(int id)
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             Category categories = _categoryRepo.FindAll().Where(c => c.ID == id).Include(p => p.Posts).FirstOrDefault();
             return View(categories);
         }
@@ -229,6 +302,8 @@ namespace Varldsklass.Web.Controllers
         /* ---- Front View ---- */
         public ActionResult CategoryList()
         {
+            if (NotAllowedHere()) return RedirectAway();
+
             List<Category> categories = _categoryRepo.FindAll().ToList();
 
             return View(categories);
